+++ src/request_error.js
@@ -1,0 +1,25 @@
+const REQUEST_ERROR_NAME = 'RequestError';
+
+/**
+ * Represents custom request error. The message is emitted as HTTP response
+ * with a specific HTTP code and headers.
+ * If this error is thrown from the `prepareRequestFunction` function,
+ * the message and status code is sent to client.
+ * By default, the response will have Content-Type: text/plain
+ * and for the 407 status the Proxy-Authenticate header will be added.
+ */
+class RequestError extends Error {
+    constructor(message, statusCode, headers) {
+        super(message);
+        this.name = REQUEST_ERROR_NAME;
+        this.statusCode = statusCode;
+        this.headers = headers;
+
+        Error.captureStackTrace(this, RequestError);
+    }
+}
+
+module.exports = {
+    RequestError,
+    REQUEST_ERROR_NAME,
+};
